# static/swagger.json
{
    "swagger": "2.0",
    "info": {
      "title": "Audiobook Store API",
      "description": "API for managing and retrieving audiobooks.",
      "version": "1.0.0"
    },
    "host": "localhost:5000",
    "basePath": "/",
    "schemes": ["http"],
    "securityDefinitions": {
      "Bearer": {
        "type": "apiKey",
        "name": "Authorization",
        "in": "header",
        "description": "JWT Authorization header using the Bearer scheme. Example: 'Bearer {token}'"
      }
    },
    "paths": {
      "/auth/register": {
      "post": {
          "tags": [
              "auth"
          ],
          "summary": "Create new user",
          "description": "Creates new user with given data.",
          "parameters": [
              {
                  "in": "body",
                  "name": "body",
                  "description": "User information to create new user",
                  "required": true,
                  "schema": {
                      "$ref": "#/definitions/User"
                  }
              }
          ],
          "responses": {
              "201": {
                  "description": "User was successfully created"
              }
          }
      }
  },
  "/auth/token": {
    "post": {
      "tags": ["auth"],
      "summary": "User login",
      "description": "Endpoint for user authentication.",
      "consumes": ["application/x-www-form-urlencoded"],
      "parameters": [
        {
          "name": "username",
          "in": "formData",
          "required": true,
          "type": "string"
        },
        {
          "name": "password",
          "in": "formData",
          "required": true,
          "type": "string",
          "format": "password"
        }
      ],
      "responses": {
        "200": {
          "description": "Login successful",
          "schema": {
            "$ref": "#/definitions/Token"
          }
        },
        "401": {
          "description": "Invalid username or password"
        }
      }
    }
  },
...
    },
    "definitions": {
      "Token": {
        "type": "object",
        "properties": {
          "access_token": {
            "type": "string"
          },
          "token_type": {
            "type": "string"
          }
        }
      },
      "User": {
        "type": "object",
        "properties": {
            "id": {
              "type": "integer",
              "format": "int64",
              "readOnly": true
            },
            "username": {
                "type": "string"
            },
            "email": {
                "type": "string"
            },
            "password": {
                "type": "string"
            },
            "role": {
                "type": "string"
            }
        }
    },
...
  }
}